Migrate variant.
