+++ crates/macros/src/common/mod.rs
@@ -1,0 +1,5 @@
+mod container;
+mod field;
+mod variant;
+
+pub use variant::*;
